Add DivIDE support
The previous release crashes when the DivIDE trap points are hit, despite us running in +3 RAM.  To avoid this we need to put the ROM in DivIDE RAM, so it's already paged.

Also corrects the RAM filling, which was setting the fill value before the MSB of the address had been calculated.