Improved tunnel trimming
The old method used a mix of black attributes and data masking to trim sprites at the screen edges, with every sprite trimmed separately.

The new method no longer uses attributes, so it's compatible with the colour version.  It also finds the extent of the trim block for all sprites, for more efficient clearing.