Fixed code size overflow crash
Adding the new keyboard code in the last release pushed the code
size beyond the 4K region allocated for it. The colour/mono
functions were then corrupted by the stack or overwritten by the
look-up tables generated at B000.

The look-up tables have been moved to B200, to give an extra 512
bytes of space. The stack has also been moved to hang back from
C000, which is after the end of the look-up tables.

Also changed the green ghost to dark yellow, to be closer to its
original orange colour. Changing the maze colour also uses the
bright version unless Shift is pressed.