Updated ReadMe file
Converted the ReadMe to Markdown format for better display on GitHub. Also
added some extra details about the emulator and its hardware requirements.

Added basic build instructions, which now mention the need for Pasmo and
tile2sam, and provide links for where to find them. We don't provide details
for each as they're too platform-specific. Thanks to Pete Moore for getting
the ball rolling on this!